net: stmmac: Fix system freeze when bridge is enabled
On i.MX8M Plus platform an issue with bridge mode is observed, causing
the whole system to freeze because register access is performed while
clocks are disabled. As a workaround, lets keep clocks enabled on this
platform.

This bug was observed with two NXP i.MX8M Plus devices:

- Phytec phyboard-POLLUX
- Advantech EPC-R3720

The BUG can be reproducible by creating a simple bridge and assign the EQOS
interface to it. During the configuration, system will freeze with no error
message. The interface works flawlessly without bridge configuration.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>